+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to the "erabasic" extension will be documented in this file.
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
+### Added
+- Multi-root Workspaces に対応
 
 ## [0.1.1] - 2017-11-05
 ### Fixed